third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/attribute-modified-ns-expected.txt

Tests that elements panel updates dom tree structure upon changing the attribute with namespace.


Running: testDumpInitial
========= Original ========
  <a id="node" xlink:href="http://localhost">link</a>

Running: testAttributeUpdated
===== On attribute set =====
  <a id="node" xlink:href="changed-url">link</a>

Running: testAttributeRemoved
=== On attribute removed ===
  <a id="node">link</a>
